+++ src/main/java/com/clone/getchu/domain/auth/dto/SignupResponse.java
@@ -2,6 +2,8 @@ package com.clone.getchu.domain.auth.dto;
 
 import com.clone.getchu.domain.member.entity.Member;
 
+import java.time.LocalDateTime;
+
 /**
  * [회원가입 응답 DTO]
  * POST /auth/signup → 201 Created
@@ -11,7 +13,8 @@ import com.clone.getchu.domain.member.entity.Member;
 public record SignupResponse(
         Long id,
         String email,
-        String nickname
+        String nickname,
+        LocalDateTime createdAt
 ) {
 
     // 엔티티 → DTO 변환 (컨트롤러/서비스 계층에서 new SignupResponse(member) 로 호출)
@@ -19,7 +22,8 @@ public record SignupResponse(
         return new SignupResponse(
                 member.getId(),
                 member.getEmail(),
-                member.getNickname()
+                member.getNickname(),
+                member.getCreatedAt()
         );
     }
 }
