+++ src/main/java/com/clone/getchu/domain/member/dto/request/UpdatePasswordRequest.java
@@ -1,0 +1,17 @@
+package com.clone.getchu.domain.member.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
+
+public record UpdatePasswordRequest(
+        @NotBlank(message = "현재 비밀번호를 입력해주세요.")
+        String oldPassword,
+
+        @NotBlank(message = "새 비밀번호를 입력해주세요.")
+        @Size(min = 8, message = "비밀번호는 8자 이상이어야 합니다.")
+
+        @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[!@#$%^&*()_+\\-=\\[\\]{};':\"\\\\|,.<>\\/?]).+$",
+                message = "비밀번호는 영문, 숫자, 특수문자를 각각 1자 이상 포함해야 합니다.")
+        String newPassword
+) {}
