+++ src/main/java/com/clone/getchu/domain/member/dto/response/MemberProfileResponse.java
@@ -3,13 +3,15 @@ package com.clone.getchu.domain.member.dto.response;
 import com.clone.getchu.domain.member.entity.Member;
 
 import java.math.BigDecimal;
+import java.time.LocalDateTime;
 
 public record MemberProfileResponse(
         Long id,
         String nickname,
         String profileImageUrl,
         BigDecimal averageRating,
-        int reviewCount
+        int reviewCount,
+        LocalDateTime createdAt
         // 리뷰 목록은 ReviewController에서 별도 API로
 ) {
     public static MemberProfileResponse from(Member member) {
@@ -18,7 +20,8 @@ public record MemberProfileResponse(
                 member.getNickname(),
                 member.getProfileImageUrl(),
                 member.getAverageRating(),
-                member.getReviewCount()
+                member.getReviewCount(),
+                member.getCreatedAt()
         );
     }
 }
