+++ src/main/java/com/clone/getchu/domain/member/dto/response/MemberResponse.java
@@ -3,6 +3,7 @@ package com.clone.getchu.domain.member.dto.response;
 import com.clone.getchu.domain.member.entity.Member;
 
 import java.math.BigDecimal;
+import java.time.LocalDateTime;
 
 public record MemberResponse(
         Long id,
@@ -10,7 +11,8 @@ public record MemberResponse(
         String nickname,
         String profileImageUrl,
         BigDecimal averageRating,
-        int reviewCount
+        int reviewCount,
+        LocalDateTime createdAt
 ) {
     public static MemberResponse from(Member member) {
         return new MemberResponse(
@@ -19,7 +21,8 @@ public record MemberResponse(
                 member.getNickname(),
                 member.getProfileImageUrl(),
                 member.getAverageRating(),
-                member.getReviewCount()
+                member.getReviewCount(),
+                member.getCreatedAt()
         );
     }
 }
