+++ src/main/java/com/clone/getchu/global/common/BaseEntity.java
@@ -17,5 +17,5 @@ public abstract class BaseEntity {
     private LocalDateTime createdAt;
 
     @LastModifiedDate
-    private LocalDateTime modifiedAt;
+    private LocalDateTime updatedAt;
 }
